packages/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "minions-match"
version = "0.1.0"
description = "Match scoring, shortlists, and hard gate evaluation between jobs and profile"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
keywords = ["match","ai","minions"]
authors = [
    { name = "Mehdi Nabhani", email = "mehdi@the-mehdi.com" }
]
dependencies = [
    "minions-sdk>=0.2.1",
]

[project.optional-dependencies]
test = ["pytest>=7.0"]

[project.urls]
Homepage = "https://github.com/mxn2020/minions-match"
Repository = "https://github.com/mxn2020/minions-match"

[tool.hatch.build.targets.wheel]
packages = ["minions_match"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
